healthcare-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0
torchvision==0.25.0
ultralytics>=8.0.0
opencv-python-headless>=4.5.0
numpy>=1.20.0
requests>=2.25.0
boto3>=1.26.0
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv>=0.19.0
huggingface_hub>=0.23.0
